refactor: move IntentsUserId type to types
