+++ src/features/account/hooks/useWatchHoldings.tsx
@@ -2,7 +2,7 @@ import { useQueries } from "@tanstack/react-query"
 import { useMemo } from "react"
 import { createTokenUsdPricesQueryOptions } from "../../../hooks/useTokensUsdPrices"
 import type { BaseTokenInfo, UnifiedTokenInfo } from "../../../types/base"
-import type { IntentsUserId } from "../../../utils/defuse"
+import type { IntentsUserId } from "../../../types/intentsUserId"
 import { getTokenId } from "../../../utils/token"
 import { getUnderlyingBaseTokenInfos } from "../../../utils/tokenUtils"
 import {
