+++ src/features/otcDesk/hooks/useOtcTakerPreparation.ts
@@ -5,8 +5,8 @@ import { logger } from "../../../logger"
 import { getDepositedBalances } from "../../../services/defuseBalanceService"
 import type { AggregatedQuote } from "../../../services/quoteService"
 import type { BaseTokenInfo, UnifiedTokenInfo } from "../../../types/base"
+import type { IntentsUserId } from "../../../types/intentsUserId"
 import { assert } from "../../../utils/assert"
-import type { IntentsUserId } from "../../../utils/defuse"
 import { isBaseToken } from "../../../utils/token"
 import { getUnderlyingBaseTokenInfos } from "../../../utils/tokenUtils"
 import {
