+++ src/features/otcDesk/stores/otcMakerTrades.ts
@@ -2,10 +2,8 @@ import { create } from "zustand"
 import { createJSONStorage, persist } from "zustand/middleware"
 import type { SignerCredentials } from "../../../core/formatters"
 import type { MultiPayload } from "../../../types/defuse-contracts-types"
-import {
-  type IntentsUserId,
-  authHandleToIntentsUserId,
-} from "../../../utils/defuse"
+import type { IntentsUserId } from "../../../types/intentsUserId"
+import { authHandleToIntentsUserId } from "../../../utils/defuse"
 
 type OtcMakerTrade = {
   tradeId: string
