+++ src/services/depositService.ts
@@ -33,9 +33,10 @@ import { logger } from "../logger"
 import { AuthMethod } from "../types/authHandle"
 import type { BaseTokenInfo, SupportedChainName } from "../types/base"
 import type { SendTransactionEVMParams, Transaction } from "../types/deposit"
+import type { IntentsUserId } from "../types/intentsUserId"
 import { BlockchainEnum } from "../types/interfaces"
 import { assert } from "../utils/assert"
-import { type IntentsUserId, authHandleToIntentsUserId } from "../utils/defuse"
+import { authHandleToIntentsUserId } from "../utils/defuse"
 import { getEVMChainId } from "../utils/evmChainId"
 import { isNativeToken } from "../utils/token"
 import { getDepositAddress, getSupportedTokens } from "./poaBridgeClient"
